code/aleksandr-k3s-highload1/requirements.txt
Flask==2.2.5
requests==2.31.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests


